% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{predict.pbc}
\alias{predict.pbc}
\title{S3 predict method for class "pbc"}
\usage{
\method{predict}{pbc}(object, Xnew, ...)
}
\arguments{
\item{object}{object of class "pbc"}

\item{Xnew}{numeric matrix of inputs}

\item{...}{additional arguments (not used)}
}
\value{
List containing predictions
}
\description{
S3 predict method for class "pbc"
}
\examples{
A <- as.matrix(expand.grid(rep(list(0:1), 6)))
set.seed(2025)
X <- A[sample(nrow(A),size=500, replace = TRUE),]
y <- as.numeric(rowSums(X)>=3)
fit <- PBclassifier(X,y)
predict(fit,X)
}
